날씨 관련 Model 직렬화 가능하도록 변경 테스트 아직 안함

+++ core/database/src/main/java/io/github/pknujsp/weatherwizard/core/database/widget/WidgetDao.kt
@@ -11,7 +11,7 @@ import kotlinx.coroutines.flow.Flow
 interface WidgetDao {
 
     @Insert(onConflict = OnConflictStrategy.REPLACE)
-    suspend fun insert(widgetDto: WidgetDto): Int
+    suspend fun insert(widgetDto: WidgetDto): Long
 
     @Query("SELECT * FROM widgets ORDER BY id DESC")
     fun getAll(): Flow<List<WidgetDto>>
